+++ src/views/SignupPage/components/SignupForm/style.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+export const formWrapper = style({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: 50,
+});
